+++ tests/test_hash.py
@@ -1,0 +1,8 @@
+"""
+The file to test the hashes made 
+"""
+import pytest
+import hash_utils
+from redis_helper_kit import * 
+
+
